Refactored library classes
