+++ static/js/components/tiles/line_tile.tsx
@@ -21,9 +21,7 @@
 import axios from "axios";
 import _ from "lodash";
 import React, { useCallback, useEffect, useRef, useState } from "react";
-import ReactDOM from "react-dom";
 
-import { DEFAULT_API_ENDPOINT } from "../../../library/constants";
 import { DataGroup, DataPoint, expandDataPoints } from "../../chart/base";
 import { drawLineChart } from "../../chart/draw";
 import { DATA_CSS_CLASS } from "../../constants/tile_constants";
@@ -38,7 +36,7 @@ import { ReplacementStrings, getStatVarName } from "../../utils/tile_utils";
 import { ChartTileContainer } from "./chart_tile";
 import { useDrawOnResize } from "./use_draw_on_resize";
 
-interface LineTilePropType {
+export interface LineTilePropType {
   id: string;
   title: string;
   place: NamedTypedPlace;
@@ -212,21 +210,3 @@ function rawToChart(
     unit,
   };
 }
-
-/**
- * Renders line chart tile component in the given HTML element
- * @param element DOM element to render the chart
- * @param props line chart tile component properties
- */
-export const renderLineComponent = (
-  element: HTMLElement,
-  props: LineTilePropType
-): void => {
-  ReactDOM.render(
-    React.createElement(LineTile, {
-      apiRoot: DEFAULT_API_ENDPOINT,
-      ...props,
-    }),
-    element
-  );
-};
