+++ static/js/components/tiles/map_tile.tsx
@@ -22,9 +22,7 @@ import axios from "axios";
 import * as d3 from "d3";
 import _ from "lodash";
 import React, { useEffect, useRef, useState } from "react";
-import ReactDOM from "react-dom";
 
-import { DEFAULT_API_ENDPOINT } from "../../../library/constants";
 import {
   addPolygonLayer,
   drawD3Map,
@@ -55,7 +53,7 @@ import { getDateRange } from "../../utils/string_utils";
 import { ReplacementStrings, getMergedSvg } from "../../utils/tile_utils";
 import { ChartTileContainer } from "./chart_tile";
 
-interface MapTilePropType {
+export interface MapTilePropType {
   id: string;
   title: string;
   place: NamedTypedPlace;
@@ -426,21 +424,3 @@ export function draw(
     );
   }
 }
-
-/**
- * Renders map chart tile component in the given HTML element
- * @param element DOM element to render the chart
- * @param props map chart tile component properties
- */
-export const renderMapComponent = (
-  element: HTMLElement,
-  props: MapTilePropType
-): void => {
-  ReactDOM.render(
-    React.createElement(MapTile, {
-      apiRoot: DEFAULT_API_ENDPOINT,
-      ...props,
-    }),
-    element
-  );
-};
